+++ .dotnet/src/Custom/FineTuning/FineTuningClient.cs
@@ -7,9 +7,11 @@ using System.Threading.Tasks;
 
 namespace OpenAI.FineTuning;
 
-/// <summary>
-/// The service client for OpenAI fine-tuning operations.
-/// </summary>
+// CUSTOM:
+// - Renamed.
+// - Suppressed constructor that takes endpoint parameter; endpoint is now a property in the options class.
+// - Suppressed convenience methods for now.
+/// <summary> The service client for OpenAI fine-tuning operations. </summary>
 [CodeGenClient("FineTuning")]
 [CodeGenSuppress("CreateFineTuningJobAsync", typeof(FineTuningOptions))]
 [CodeGenSuppress("CreateFineTuningJob", typeof(FineTuningOptions))]
@@ -25,20 +27,28 @@ namespace OpenAI.FineTuning;
 [CodeGenSuppress("GetFineTuningJobCheckpoints", typeof(string), typeof(string), typeof(int?))]
 public partial class FineTuningClient
 {
-    // Customization: documented constructors, apply protected visibility    
+    // CUSTOM:
+    // - Used a custom pipeline.
+    // - Demoted the endpoint parameter to be a property in the options class.
+    /// <summary> Initializes a new instance of <see cref="FineTuningClient">. </summary>
+    /// <param name="credential"> The API key to authenticate with the service. </param>
+    /// <exception cref="ArgumentNullException"> <paramref name="credential"/> is null. </exception>
+    public FineTuningClient(ApiKeyCredential credential) : this(credential, new OpenAIClientOptions())
+    {
+    }
 
-    /// <summary>
-    /// Initializes a new instance of <see cref="FineTuningClient"/> that will use an API key when authenticating.
-    /// </summary>
-    /// <param name="credential"> The API key used to authenticate with the service endpoint. </param>
-    /// <param name="options"> Additional options to customize the client. </param>
-    /// <exception cref="ArgumentNullException"> The provided <paramref name="credential"/> was null. </exception>
-    public FineTuningClient(ApiKeyCredential credential, OpenAIClientOptions options = null)
-        : this(
-              OpenAIClient.CreatePipeline(OpenAIClient.GetApiKey(credential, requireExplicitCredential: true), options),
-              OpenAIClient.GetEndpoint(options),
-              options)
-    { }
+    // Customization: documented constructors, apply protected visibility    
+    // CUSTOM:
+    // - Used a custom pipeline.
+    // - Demoted the endpoint parameter to be a property in the options class.
+    /// <summary> Initializes a new instance of <see cref="FineTuningClient">. </summary>
+    /// <param name="credential"> The API key to authenticate with the service. </param>
+    /// <param name="options"> The options to configure the client. </param>
+    /// <exception cref="ArgumentNullException"> <paramref name="credential"/> is null. </exception>
+    public FineTuningClient(ApiKeyCredential credential, OpenAIClientOptions options)
+    {
+        Argument.AssertNotNull(credential, nameof(credential));
+        options ??= new OpenAIClientOptions();
 
     /// <summary>
     /// Initializes a new instance of <see cref="FineTuningClient"/> that will use an API key from the OPENAI_API_KEY
@@ -57,13 +67,21 @@ public partial class FineTuningClient
               options)
     { }
 
-    /// <summary> Initializes a new instance of FineTuningClient. </summary>
-    /// <param name="pipeline"> The HTTP pipeline for sending and receiving REST requests and responses. </param>
-    /// <param name="endpoint"> OpenAI Endpoint. </param>
-    protected internal FineTuningClient(ClientPipeline pipeline, Uri endpoint, OpenAIClientOptions options)
+    // CUSTOM:
+    // - Used a custom pipeline.
+    // - Demoted the endpoint parameter to be a property in the options class.
+    // - Made protected.
+    /// <summary> Initializes a new instance of <see cref="FineTuningClient">. </summary>
+    /// <param name="pipeline"> The HTTP pipeline to send and receive REST requests and responses. </param>
+    /// <param name="options"> The options to configure the client. </param>
+    /// <exception cref="ArgumentNullException"> <paramref name="pipeline"/> is null. </exception>
+    protected internal FineTuningClient(ClientPipeline pipeline, OpenAIClientOptions options)
     {
+        Argument.AssertNotNull(pipeline, nameof(pipeline));
+        options ??= new OpenAIClientOptions();
+
         _pipeline = pipeline;
-        _endpoint = endpoint;
+        _endpoint = OpenAIClient.GetEndpoint(options);
     }
 
     /// <summary> Creates a job with a training file and model. </summary>