+++ .dotnet/tests/Batch/BatchTests.cs
@@ -60,8 +60,8 @@ public partial class BatchTests : SyncAsyncTestBase
         streamWriter.Flush();
         testFileStream.Position = 0;
 
-        FileClient fileClient = new();
-        OpenAIFileInfo inputFile = await fileClient.UploadFileAsync(testFileStream, "test-batch-sampleFile", FileUploadPurpose.Batch);
+        FileClient fileClient = GetTestClient<FileClient>(TestScenario.Files);
+        OpenAIFileInfo inputFile = await fileClient.UploadFileAsync(testFileStream, "test-batch-file", FileUploadPurpose.Batch);
         Assert.That(inputFile.Id, Is.Not.Null.And.Not.Empty);
 
         BatchClient client = GetTestClient();