+++ .scripts/Invoke-CodeGen.ps1
@@ -15,7 +15,7 @@ try {
   Invoke { npm exec --no -- tsp format **/*tsp }
   Invoke { npm exec --no -- tsp compile . --pretty }
   Invoke { .$PSScriptRoot\Update-ClientModel.ps1 }
-  Invoke { .$PSScriptRoot\Edit-Serialization.ps1 }
+  Invoke { .$PSScriptRoot\Edit-Deserialization.ps1 }
   Invoke { .$PSScriptRoot\Run-Checks.ps1 }
 }
 finally {